Move biome change listener to LevelSpawnSystem
